Rename p2p_types -> types
