+++ elements/lisk-p2p/src/index.ts
@@ -17,6 +17,6 @@ import * as errors from './errors';
 import * as events from './events';
 import { P2P } from './p2p';
 import * as p2p_request from './p2p_request';
-import * as p2p_types from './p2p_types';
+import * as p2p_types from './types';
 
 export { constants, errors, events, p2p_request, p2p_types, P2P };
