+++ elements/lisk-p2p/src/peer/inbound.ts
@@ -27,7 +27,7 @@ import {
 	REMOTE_SC_EVENT_MESSAGE,
 	REMOTE_SC_EVENT_RPC_REQUEST,
 } from '../events';
-import { P2PPeerInfo } from '../p2p_types';
+import { P2PPeerInfo } from '../types';
 
 import {
 	Peer,
