+++ elements/lisk-p2p/src/peer/outbound.ts
@@ -38,7 +38,7 @@ import {
 	P2PPeerInfo,
 	P2PRequestPacket,
 	P2PResponsePacket,
-} from '../p2p_types';
+} from '../types';
 
 import {
 	Peer,
