+++ elements/lisk-p2p/src/utils/select.ts
@@ -21,7 +21,7 @@ import {
 	P2PPeerSelectionForConnectionInput,
 	P2PPeerSelectionForRequestInput,
 	P2PPeerSelectionForSendInput,
-} from '../p2p_types';
+} from '../types';
 
 const _removeCommonIPsFromLists = (
 	peerList: ReadonlyArray<P2PPeerInfo>,
